moved selectitem to it's own file
